Add new options to base app from thesis
Former-commit-id: 2f3eef1d23273ca03d405a44a345217f85fbf2ea

+++ myna/application/exaca/microstructure_region/configure.py
@@ -19,7 +19,8 @@ def setup_case(
 ):
 
     # Copy template to case directory
-    shutil.copytree(sim.args.template, case_dir, dirs_exist_ok=True)
+    if (not os.path.exists(sim.args.template)) or (sim.args.overwrite):
+        shutil.copytree(sim.args.template, case_dir, dirs_exist_ok=True)
 
     # Get case settings and template input JSON
     myna_settings = load_input(os.path.join(case_dir, "myna_data.yaml"))
